requirements.txt
python-telegram-bot==13.4.1

psycopg2==2.8.6